feat(plugin): add PluginStatus component for managing plugin status display and updates
