+++ .vscode/settings.json
@@ -40,6 +40,7 @@
     "gapis",
     "hangingpiece",
     "hideloading",
+    "hilightjs",
     "hintrc",
     "hljs",
     "iconsrc",
